ARTEMIS-2316 Reverting change on AMQPMessage, and changing test to be accurate
This (kind of) reverts commit bd1162d9b8b3563437dd4fc2dafa0fd4e56cd8df.
